move back code to NextGen
